+++ frontend/.gitignore
@@ -209,3 +209,6 @@ ehthumbs.db
 Thumbs.db
 
 ./firebase
+/firebase-debug.log
+/firebase-debug.*.log
+.firebaserc
