+++ frontend/src/components/features/CategoryButton.vue
