+++ frontend/src/components/features/CategoryTabs.vue
