+++ frontend/src/components/modals/EmailModal.vue
